test/requirements.txt
Flask==2.0.2
flask-cors==3.1.1
mysql-connector-python==8.0.28
